All checks pass. The implementation is complete with the linter issues fixed by moving state and callbacks to the container component.
